requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorboardX>=2.6
gymnasium[classic-control]>=0.29.0
numpy>=1.21
pyyaml>=6.0
